rename file one shot player
