+++ src/lib.rs
@@ -1,5 +1,5 @@
 mod audio_player;
-mod play_sound_event;
+mod one_shot_player;
 mod plugin;
 mod sync_wrapped;
 mod utils;
@@ -7,6 +7,6 @@ mod utils;
 pub use crate::audio_player::AudioListener;
 pub use crate::audio_player::AudioSource;
 pub use crate::audio_player::AudioSourcePlayer;
-pub use crate::play_sound_event::OneShotPlayer;
+pub use crate::one_shot_player::OneShotPlayer;
 pub use crate::plugin::FmodPlugin;
 pub use crate::sync_wrapped::event_instance::EventInstance;
