+++ src/one_shot_player.rs
